helpers: move out memoized_property
