+++ tests/test_disassemble_helpers.py
@@ -18,7 +18,6 @@ from strcs.disassemble import (
     fields_from_attrs,
     fields_from_class,
     fields_from_dataclasses,
-    memoized_property,
 )
 
 
@@ -476,73 +475,6 @@ describe "fields_from_dataclasses":
         assert thing._thing == "one"
         assert thing.other == 4
 
-
-describe "memoized_property":
-    it "memoizes":
-        called: list[int] = []
-
-        class Thing:
-            @memoized_property
-            def blah(self) -> str:
-                called.append(1)
-                return "stuff"
-
-        thing = Thing()
-        assert thing.blah == "stuff"
-        assert called == [1]
-        assert thing.blah == "stuff"
-        assert called == [1]
-        assert thing.blah == "stuff"
-        assert called == [1]
-
-    it "allows setting the value":
-        called: list[int] = []
-
-        class Thing:
-            @memoized_property
-            def blah(self) -> str:
-                called.append(1)
-                return "stuff"
-
-        thing = Thing()
-        assert thing.blah == "stuff"
-        assert called == [1]
-        thing.blah = "other"
-        assert thing.blah == "other"
-        assert called == [1]
-        assert thing.blah == "other"
-        assert called == [1]
-
-    it "allows deleting the value":
-        called: list[int] = []
-
-        class Thing:
-            @memoized_property
-            def blah(self) -> str:
-                called.append(1)
-                return "stuff"
-
-        thing = Thing()
-        assert thing.blah == "stuff"
-        assert called == [1]
-        del thing.blah
-        assert thing.blah == "stuff"
-        assert called == [1, 1]
-        assert thing.blah == "stuff"
-        assert called == [1, 1]
-
-    it "keeps the type annotation":
-
-        class Thing:
-            @memoized_property
-            def blah(self) -> str:
-                return "stuff"
-
-        def a(things: str) -> None:
-            pass
-
-        # will make mypy complain if it's broken
-        a(Thing().blah)
 
 describe "IsAnnotated":
     it "knows if something is annotated":
